fix: fix click on card
