+++ src/components/CardsList/index.tsx
@@ -1,6 +1,5 @@
 import styles from './CardList.module.css';
 import CardItem from '../CardItem';
-import { useNavigate, useParams } from 'react-router-dom';
 import { Hero } from '../../types/types';
 
 export type CardsProps = {
@@ -8,23 +7,12 @@ export type CardsProps = {
 };
 
 const CardsList: React.FC<CardsProps> = ({ heroes }) => {
-  const navigate = useNavigate();
-  const { page } = useParams<{ page: string }>();
-
   return (
     <div className={styles.wrapper}>
       <h2>Results Found :</h2>
       <div className={styles.cardsList}>
         {heroes.map((hero, index) => (
-          <div
-            onClick={() => {
-              const linkArray = hero.url.split('/');
-              const userIndex = linkArray[linkArray.length - 2];
-              navigate(`/search/${page}/details/${userIndex}`);
-            }}
-          >
-            <CardItem key={index} hero={hero} />
-          </div>
+          <CardItem key={index} hero={hero} />
         ))}
       </div>
     </div>
