+++ src/components/Flyout/Flyout.module.css
@@ -1,7 +1,7 @@
 .flyout {
     position: fixed;
     bottom: 20px;
-    right: 10%;
+    right: 5%;
     transform: translateX(-50%);
     background-color: #f8f8f8;
     padding: 10px 20px;
